refactor profile into personal and users
