+++ src/Core/Application/Identity/Users/IUserService.cs
@@ -1,3 +1,6 @@
+using System.Security.Claims;
+using FSH.WebApi.Application.Identity.Users.Password;
+
 namespace FSH.WebApi.Application.Identity.Users;
 
 public interface IUserService : ITransientService
@@ -15,7 +18,6 @@ public interface IUserService : ITransientService
     Task<UserDetailsDto> GetAsync(string userId, CancellationToken cancellationToken);
 
     Task<List<UserRoleDto>> GetRolesAsync(string userId, CancellationToken cancellationToken);
-
     Task<string> AssignRolesAsync(string userId, UserRolesRequest request, CancellationToken cancellationToken);
 
     Task<List<string>> GetPermissionsAsync(string userId, CancellationToken cancellationToken);
@@ -23,4 +25,15 @@ public interface IUserService : ITransientService
     Task InvalidatePermissionCacheAsync(string userId, CancellationToken cancellationToken);
 
     Task ToggleStatusAsync(ToggleUserStatusRequest request, CancellationToken cancellationToken);
+
+    Task<string> GetOrCreateFromPrincipalAsync(ClaimsPrincipal principal);
+    Task<string> CreateAsync(CreateUserRequest request, string origin);
+    Task UpdateAsync(UpdateUserRequest request, string userId);
+
+    Task<string> ConfirmEmailAsync(string userId, string code, string tenant, CancellationToken cancellationToken);
+    Task<string> ConfirmPhoneNumberAsync(string userId, string code);
+
+    Task<string> ForgotPasswordAsync(ForgotPasswordRequest request, string origin);
+    Task<string> ResetPasswordAsync(ResetPasswordRequest request);
+    Task ChangePasswordAsync(ChangePasswordRequest request, string userId);
 }
