+++ src/Core/Application/Identity/Users/Password/ChangePasswordRequest.cs
@@ -1,4 +1,4 @@
-namespace FSH.WebApi.Application.Identity.Users.Profile;
+namespace FSH.WebApi.Application.Identity.Users.Password;
 
 public class ChangePasswordRequest
 {
