+++ src/Host/Controllers/Identity/UsersController.cs
@@ -1,4 +1,5 @@
 using FSH.WebApi.Application.Identity.Users;
+using FSH.WebApi.Application.Identity.Users.Password;
 
 namespace FSH.WebApi.Host.Controllers.Identity;
 
@@ -18,7 +19,7 @@ public class UsersController : VersionNeutralApiController
 
     [HttpGet("{id}")]
     [MustHavePermission(FSHAction.View, FSHResource.Users)]
-    [OpenApiOperation("Get user details.", "")]
+    [OpenApiOperation("Get a user's details.", "")]
     public Task<UserDetailsDto> GetByIdAsync(string id, CancellationToken cancellationToken)
     {
         return _userService.GetAsync(id, cancellationToken);
@@ -35,17 +36,28 @@ public class UsersController : VersionNeutralApiController
     [HttpPost("{id}/roles")]
     [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Register))]
     [MustHavePermission(FSHAction.Update, FSHResource.UserRoles)]
-    [OpenApiOperation("Change a user's assigned roles.", "")]
+    [OpenApiOperation("Update a user's assigned roles.", "")]
     public Task<string> AssignRolesAsync(string id, UserRolesRequest request, CancellationToken cancellationToken)
     {
         return _userService.AssignRolesAsync(id, request, cancellationToken);
+    }
+
+    [HttpPost]
+    [AllowAnonymous]
+    [OpenApiOperation("Create a new user.", "")]
+    public Task<string> CreateAsync(CreateUserRequest request)
+    {
+        // TODO: check if registering anonymous users is actually allowed (should probably be an appsetting)
+        // and return UnAuthorized when it isn't
+        // Also: add other protection to prevent automatic posting (captcha?)
+        return _userService.CreateAsync(request, GetOriginFromRequest());
     }
 
     [HttpPost("{id}/toggle-status")]
     [MustHavePermission(FSHAction.Update, FSHResource.Users)]
     [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Register))]
     [OpenApiOperation("Toggle a user's active status.", "")]
-    public async Task<ActionResult> ToggleUserStatusAsync(string id, ToggleUserStatusRequest request, CancellationToken cancellationToken)
+    public async Task<ActionResult> ToggleStatusAsync(string id, ToggleUserStatusRequest request, CancellationToken cancellationToken)
     {
         if (id != request.UserId)
         {
@@ -55,4 +67,42 @@ public class UsersController : VersionNeutralApiController
         await _userService.ToggleStatusAsync(request, cancellationToken);
         return Ok();
     }
+
+    [HttpGet("confirm-email")]
+    [AllowAnonymous]
+    [OpenApiOperation("Confirm email address for a user.", "")]
+    [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Search))]
+    public Task<string> ConfirmEmailAsync([FromQuery] string tenant, [FromQuery] string userId, [FromQuery] string code, CancellationToken cancellationToken)
+    {
+        return _userService.ConfirmEmailAsync(userId, code, tenant, cancellationToken);
+    }
+
+    [HttpGet("confirm-phone-number")]
+    [AllowAnonymous]
+    [OpenApiOperation("Confirm phone number for a user.", "")]
+    [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Search))]
+    public Task<string> ConfirmPhoneNumberAsync([FromQuery] string userId, [FromQuery] string code)
+    {
+        return _userService.ConfirmPhoneNumberAsync(userId, code);
+    }
+
+    [HttpPost("forgot-password")]
+    [AllowAnonymous]
+    [TenantIdHeader]
+    [OpenApiOperation("Request a pasword reset email for a user.", "")]
+    [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Register))]
+    public Task<string> ForgotPasswordAsync(ForgotPasswordRequest request)
+    {
+        return _userService.ForgotPasswordAsync(request, GetOriginFromRequest());
+    }
+
+    [HttpPost("reset-password")]
+    [OpenApiOperation("Reset a user's password.", "")]
+    [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Register))]
+    public Task<string> ResetPasswordAsync(ResetPasswordRequest request)
+    {
+        return _userService.ResetPasswordAsync(request);
+    }
+
+    private string GetOriginFromRequest() => $"{Request.Scheme}://{Request.Host.Value}{Request.PathBase.Value}";
 }
