+++ src/Host/Controllers/Personal/PersonalController.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using FSH.WebApi.Application.Auditing;
 using FSH.WebApi.Application.Identity.Users;
+using FSH.WebApi.Application.Identity.Users.Password;
 
 namespace FSH.WebApi.Host.Controllers.Identity;
 
@@ -10,22 +11,55 @@ public class PersonalController : VersionNeutralApiController
 
     public PersonalController(IUserService userService) => _userService = userService;
 
-    [HttpGet("logs")]
-    [OpenApiOperation("Get audit logs of currently logged in user.", "")]
-    public Task<List<AuditDto>> GetMyLogsAsync()
+    [HttpGet("profile")]
+    [OpenApiOperation("Get profile details of currently logged in user.", "")]
+    public async Task<ActionResult<UserDetailsDto>> GetProfileAsync(CancellationToken cancellationToken)
     {
-        return Mediator.Send(new GetMyAuditLogsRequest());
+        return User.GetUserId() is not { } userId || string.IsNullOrEmpty(userId)
+            ? Unauthorized()
+            : Ok(await _userService.GetAsync(userId, cancellationToken));
     }
 
-    [HttpGet("permissions")]
-    [OpenApiOperation("Get permissions of currently logged in user.", "")]
-    public async Task<ActionResult<List<string>>> GetMyPermissionsAsync(CancellationToken cancellationToken)
+    [HttpPut("profile")]
+    [OpenApiOperation("Update profile details of currently logged in user.", "")]
+    public async Task<ActionResult> UpdateProfileAsync(UpdateUserRequest request)
     {
         if (User.GetUserId() is not { } userId || string.IsNullOrEmpty(userId))
         {
             return Unauthorized();
         }
 
-        return Ok(await _userService.GetPermissionsAsync(userId, cancellationToken));
+        await _userService.UpdateAsync(request, userId);
+        return Ok();
+    }
+
+    [HttpPut("change-password")]
+    [OpenApiOperation("Change password of currently logged in user.", "")]
+    [ApiConventionMethod(typeof(FSHApiConventions), nameof(FSHApiConventions.Register))]
+    public async Task<ActionResult> ChangePasswordAsync(ChangePasswordRequest model)
+    {
+        if (User.GetUserId() is not { } userId || string.IsNullOrEmpty(userId))
+        {
+            return Unauthorized();
+        }
+
+        await _userService.ChangePasswordAsync(model, userId);
+        return Ok();
+    }
+
+    [HttpGet("permissions")]
+    [OpenApiOperation("Get permissions of currently logged in user.", "")]
+    public async Task<ActionResult<List<string>>> GetPermissionsAsync(CancellationToken cancellationToken)
+    {
+        return User.GetUserId() is not { } userId || string.IsNullOrEmpty(userId)
+            ? Unauthorized()
+            : Ok(await _userService.GetPermissionsAsync(userId, cancellationToken));
+    }
+
+    [HttpGet("logs")]
+    [OpenApiOperation("Get audit logs of currently logged in user.", "")]
+    public Task<List<AuditDto>> GetLogsAsync()
+    {
+        return Mediator.Send(new GetMyAuditLogsRequest());
     }
 }
