--- src/Infrastructure/Identity/ProfileService.cs
@@ -1,325 +1,0 @@
-using System.Security.Claims;
-using System.Text;
-using Finbuckle.MultiTenant;
-using FSH.WebApi.Application.Common.Events;
-using FSH.WebApi.Application.Common.Exceptions;
-using FSH.WebApi.Application.Common.FileStorage;
-using FSH.WebApi.Application.Common.Interfaces;
-using FSH.WebApi.Application.Common.Mailing;
-using FSH.WebApi.Application.Identity;
-using FSH.WebApi.Application.Identity.Users.Profile;
-using FSH.WebApi.Domain.Common;
-using FSH.WebApi.Domain.Identity;
-using FSH.WebApi.Infrastructure.Common;
-using FSH.WebApi.Infrastructure.Mailing;
-using FSH.WebApi.Shared.Authorization;
-using FSH.WebApi.Shared.Multitenancy;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Localization;
-using Microsoft.Extensions.Options;
-using Microsoft.Identity.Web;
-
-namespace FSH.WebApi.Infrastructure.Identity;
-
-internal class ProfileService : IProfileService
-{
-    private readonly SignInManager<ApplicationUser> _signInManager;
-    private readonly UserManager<ApplicationUser> _userManager;
-    private readonly RoleManager<ApplicationRole> _roleManager;
-    private readonly IStringLocalizer<ProfileService> _localizer;
-    private readonly IJobService _jobService;
-    private readonly IMailService _mailService;
-    private readonly MailSettings _mailSettings;
-    private readonly IEmailTemplateService _templateService;
-    private readonly IFileStorageService _fileStorage;
-    private readonly IEventService _eventService;
-    private readonly ITenantInfo _currentTenant;
-
-    public ProfileService(
-        SignInManager<ApplicationUser> signInManager,
-        UserManager<ApplicationUser> userManager,
-        RoleManager<ApplicationRole> roleManager,
-        IStringLocalizer<ProfileService> localizer,
-        IJobService jobService,
-        IMailService mailService,
-        IOptions<MailSettings> mailSettings,
-        IEmailTemplateService templateService,
-        IFileStorageService fileStorage,
-        IEventService eventService,
-        ITenantInfo currentTenant)
-    {
-        _signInManager = signInManager;
-        _userManager = userManager;
-        _roleManager = roleManager;
-        _localizer = localizer;
-        _jobService = jobService;
-        _mailService = mailService;
-        _mailSettings = mailSettings.Value;
-        _templateService = templateService;
-        _fileStorage = fileStorage;
-        _currentTenant = currentTenant;
-        _eventService = eventService;
-    }
-
-    /// <summary>
-    /// This is used when authenticating with AzureAd.
-    /// The local user is retrieved using the objectidentifier claim present in the ClaimsPrincipal.
-    /// If no such claim is found, an InternalServerException is thrown.
-    /// If no user is found with that ObjectId, a new one is created and populated with the values from the ClaimsPrincipal.
-    /// If a role claim is present in the principal, and the user is not yet in that roll, then the user is added to that role.
-    /// </summary>
-    public async Task<string> GetOrCreateFromPrincipalAsync(ClaimsPrincipal principal)
-    {
-        string? objectId = principal.GetObjectId();
-        if (string.IsNullOrWhiteSpace(objectId))
-        {
-            throw new InternalServerException(_localizer["Invalid objectId"]);
-        }
-
-        var user = await _userManager.Users.Where(u => u.ObjectId == objectId).FirstOrDefaultAsync()
-            ?? await CreateOrUpdateFromPrincipalAsync(principal);
-
-        if (principal.FindFirstValue(ClaimTypes.Role) is string role &&
-            await _roleManager.RoleExistsAsync(role) &&
-            !await _userManager.IsInRoleAsync(user, role))
-        {
-            await _userManager.AddToRoleAsync(user, role);
-        }
-
-        return user.Id;
-    }
-
-    private async Task<ApplicationUser> CreateOrUpdateFromPrincipalAsync(ClaimsPrincipal principal)
-    {
-        string? email = principal.FindFirstValue(ClaimTypes.Upn);
-        string? username = principal.GetDisplayName();
-        if (string.IsNullOrEmpty(email) || string.IsNullOrEmpty(username))
-        {
-            throw new InternalServerException(string.Format(_localizer["Username or Email not valid."]));
-        }
-
-        var user = await _userManager.FindByNameAsync(username);
-        if (user is not null && !string.IsNullOrWhiteSpace(user.ObjectId))
-        {
-            throw new InternalServerException(string.Format(_localizer["Username {0} is already taken."], username));
-        }
-
-        if (user is null)
-        {
-            user = await _userManager.FindByEmailAsync(email);
-            if (user is not null && !string.IsNullOrWhiteSpace(user.ObjectId))
-            {
-                throw new InternalServerException(string.Format(_localizer["Email {0} is already taken."], email));
-            }
-        }
-
-        IdentityResult? result;
-        if (user is not null)
-        {
-            user.ObjectId = principal.GetObjectId();
-            result = await _userManager.UpdateAsync(user);
-
-            await _eventService.PublishAsync(new ApplicationUserUpdatedEvent(user.Id));
-        }
-        else
-        {
-            user = new ApplicationUser
-            {
-                ObjectId = principal.GetObjectId(),
-                FirstName = principal.FindFirstValue(ClaimTypes.GivenName),
-                LastName = principal.FindFirstValue(ClaimTypes.Surname),
-                Email = email,
-                NormalizedEmail = email.ToUpperInvariant(),
-                UserName = username,
-                NormalizedUserName = username.ToUpperInvariant(),
-                EmailConfirmed = true,
-                PhoneNumberConfirmed = true,
-                IsActive = true
-            };
-            result = await _userManager.CreateAsync(user);
-
-            await _eventService.PublishAsync(new ApplicationUserCreatedEvent(user.Id));
-        }
-
-        if (!result.Succeeded)
-        {
-            throw new InternalServerException(_localizer["Validation Errors Occurred."], result.GetErrors(_localizer));
-        }
-
-        return user;
-    }
-
-    public async Task<string> CreateAsync(CreateProfileRequest request, string origin)
-    {
-        var user = new ApplicationUser
-        {
-            Email = request.Email,
-            FirstName = request.FirstName,
-            LastName = request.LastName,
-            UserName = request.UserName,
-            PhoneNumber = request.PhoneNumber,
-            IsActive = true
-        };
-
-        var result = await _userManager.CreateAsync(user, request.Password);
-        if (!result.Succeeded)
-        {
-            throw new InternalServerException(_localizer["Validation Errors Occurred."], result.GetErrors(_localizer));
-        }
-
-        await _userManager.AddToRoleAsync(user, FSHRoles.Basic);
-
-        var messages = new List<string> { string.Format(_localizer["User {0} Registered."], user.UserName) };
-
-        if (_mailSettings.EnableVerification && !string.IsNullOrEmpty(user.Email))
-        {
-            // send verification email
-            string emailVerificationUri = await GetEmailVerificationUriAsync(user, origin);
-            var mailRequest = new MailRequest(
-                new List<string> { user.Email },
-                _localizer["Confirm Registration"],
-                _templateService.GenerateEmailConfirmationMail(user.UserName ?? "User", user.Email, emailVerificationUri));
-            _jobService.Enqueue(() => _mailService.SendAsync(mailRequest));
-            messages.Add(_localizer[$"Please check {user.Email} to verify your account!"]);
-        }
-
-        await _eventService.PublishAsync(new ApplicationUserCreatedEvent(user.Id));
-
-        return string.Join(Environment.NewLine, messages);
-    }
-
-    public async Task UpdateAsync(UpdateProfileRequest request, string userId)
-    {
-        var user = await _userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
-
-        string currentImage = user.ImageUrl ?? string.Empty;
-        if (request.Image != null || request.DeleteCurrentImage)
-        {
-            user.ImageUrl = await _fileStorage.UploadAsync<ApplicationUser>(request.Image, FileType.Image);
-            if (request.DeleteCurrentImage && !string.IsNullOrEmpty(currentImage))
-            {
-                string root = Directory.GetCurrentDirectory();
-                _fileStorage.Remove(Path.Combine(root, currentImage));
-            }
-        }
-
-        user.FirstName = request.FirstName;
-        user.LastName = request.LastName;
-        user.PhoneNumber = request.PhoneNumber;
-        string phoneNumber = await _userManager.GetPhoneNumberAsync(user);
-        if (request.PhoneNumber != phoneNumber)
-        {
-            await _userManager.SetPhoneNumberAsync(user, request.PhoneNumber);
-        }
-
-        var result = await _userManager.UpdateAsync(user);
-
-        await _signInManager.RefreshSignInAsync(user);
-
-        await _eventService.PublishAsync(new ApplicationUserUpdatedEvent(user.Id));
-
-        if (!result.Succeeded)
-        {
-            throw new InternalServerException(_localizer["Update profile failed"], result.GetErrors(_localizer));
-        }
-    }
-
-    private async Task<string> GetEmailVerificationUriAsync(ApplicationUser user, string origin)
-    {
-        string code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-        code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-        const string route = "api/profile/confirm-email/";
-        var endpointUri = new Uri(string.Concat($"{origin}/", route));
-        string verificationUri = QueryHelpers.AddQueryString(endpointUri.ToString(), QueryStringKeys.UserId, user.Id);
-        verificationUri = QueryHelpers.AddQueryString(verificationUri, QueryStringKeys.Code, code);
-        verificationUri = QueryHelpers.AddQueryString(verificationUri, MultitenancyConstants.TenantIdName, _currentTenant.Id);
-        return verificationUri;
-    }
-
-    public async Task<string> ConfirmEmailAsync(string userId, string code, string tenant, CancellationToken cancellationToken)
-    {
-        var user = await _userManager.Users
-            .Where(u => u.Id == userId && !u.EmailConfirmed)
-            .FirstOrDefaultAsync(cancellationToken);
-
-        _ = user ?? throw new InternalServerException(_localizer["An error occurred while confirming E-Mail."]);
-
-        code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
-        var result = await _userManager.ConfirmEmailAsync(user, code);
-
-        return result.Succeeded
-            ? string.Format(_localizer["Account Confirmed for E-Mail {0}. You can now use the /api/tokens endpoint to generate JWT."], user.Email)
-            : throw new InternalServerException(string.Format(_localizer["An error occurred while confirming {0}"], user.Email));
-    }
-
-    public async Task<string> ConfirmPhoneNumberAsync(string userId, string code)
-    {
-        var user = await _userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new InternalServerException(_localizer["An error occurred while confirming Mobile Phone."]);
-
-        var result = await _userManager.ChangePhoneNumberAsync(user, user.PhoneNumber, code);
-
-        return result.Succeeded
-            ? user.EmailConfirmed
-                ? string.Format(_localizer["Account Confirmed for Phone Number {0}. You can now use the /api/tokens endpoint to generate JWT."], user.PhoneNumber)
-                : string.Format(_localizer["Account Confirmed for Phone Number {0}. You should confirm your E-mail before using the /api/tokens endpoint to generate JWT."], user.PhoneNumber)
-            : throw new InternalServerException(string.Format(_localizer["An error occurred while confirming {0}"], user.PhoneNumber));
-    }
-
-    public async Task<string> ForgotPasswordAsync(ForgotPasswordRequest request, string origin)
-    {
-        var user = await _userManager.FindByEmailAsync(request.Email.Normalize());
-        if (user is null || !await _userManager.IsEmailConfirmedAsync(user))
-        {
-            // Don't reveal that the user does not exist or is not confirmed
-            throw new InternalServerException(_localizer["An Error has occurred!"]);
-        }
-
-        // For more information on how to enable account confirmation and password reset please
-        // visit https://go.microsoft.com/fwlink/?LinkID=532713
-        string code = await _userManager.GeneratePasswordResetTokenAsync(user);
-        const string route = "account/reset-password";
-        var endpointUri = new Uri(string.Concat($"{origin}/", route));
-        string passwordResetUrl = QueryHelpers.AddQueryString(endpointUri.ToString(), "Token", code);
-        var mailRequest = new MailRequest(
-            new List<string> { request.Email },
-            _localizer["Reset Password"],
-            _localizer[$"Your Password Reset Token is '{code}'. You can reset your password using the {endpointUri} Endpoint."]);
-        _jobService.Enqueue(() => _mailService.SendAsync(mailRequest));
-
-        return _localizer["Password Reset Mail has been sent to your authorized Email."];
-    }
-
-    public async Task<string> ResetPasswordAsync(ResetPasswordRequest request)
-    {
-        var user = await _userManager.FindByEmailAsync(request.Email?.Normalize());
-
-        // Don't reveal that the user does not exist
-        _ = user ?? throw new InternalServerException(_localizer["An Error has occurred!"]);
-
-        var result = await _userManager.ResetPasswordAsync(user, request.Token, request.Password);
-
-        return result.Succeeded
-            ? _localizer["Password Reset Successful!"]
-            : throw new InternalServerException(_localizer["An Error has occurred!"]);
-    }
-
-    public async Task ChangePasswordAsync(ChangePasswordRequest model, string userId)
-    {
-        var user = await _userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
-
-        var result = await _userManager.ChangePasswordAsync(user, model.Password, model.NewPassword);
-
-        if (!result.Succeeded)
-        {
-            throw new InternalServerException(_localizer["Change password failed"], result.GetErrors(_localizer));
-        }
-    }
-}
