+++ src/Infrastructure/Identity/UserService.Permissions.cs
@@ -1,0 +1,43 @@
+﻿using FSH.WebApi.Application.Common.Caching;
+using FSH.WebApi.Application.Common.Exceptions;
+using FSH.WebApi.Shared.Authorization;
+using Microsoft.EntityFrameworkCore;
+
+namespace FSH.WebApi.Infrastructure.Identity;
+
+internal partial class UserService
+{
+    public async Task<List<string>> GetPermissionsAsync(string userId, CancellationToken cancellationToken)
+    {
+        var user = await _userManager.FindByIdAsync(userId);
+
+        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
+
+        var userRoles = await _userManager.GetRolesAsync(user);
+        var permissions = new List<string>();
+        foreach (var role in await _roleManager.Roles
+            .Where(r => userRoles.Contains(r.Name))
+            .ToListAsync(cancellationToken))
+        {
+            permissions.AddRange(await _db.RoleClaims
+                .Where(rc => rc.RoleId == role.Id && rc.ClaimType == FSHClaims.Permission)
+                .Select(rc => rc.ClaimValue)
+                .ToListAsync(cancellationToken));
+        }
+
+        return permissions.Distinct().ToList();
+    }
+
+    public async Task<bool> HasPermissionAsync(string userId, string permission, CancellationToken cancellationToken)
+    {
+        var permissions = await _cache.GetOrSetAsync(
+            _cacheKeys.GetCacheKey(FSHClaims.Permission, userId),
+            () => GetPermissionsAsync(userId, cancellationToken),
+            cancellationToken: cancellationToken);
+
+        return permissions?.Contains(permission) ?? false;
+    }
+
+    public Task InvalidatePermissionCacheAsync(string userId, CancellationToken cancellationToken) =>
+        _cache.RemoveAsync(_cacheKeys.GetCacheKey(FSHClaims.Permission, userId), cancellationToken);
+}
