+++ src/Infrastructure/Identity/UserService.Roles.cs
@@ -1,0 +1,85 @@
+﻿using FSH.WebApi.Application.Common.Exceptions;
+using FSH.WebApi.Application.Identity.Users;
+using FSH.WebApi.Domain.Identity;
+using FSH.WebApi.Shared.Authorization;
+using FSH.WebApi.Shared.Multitenancy;
+using Microsoft.EntityFrameworkCore;
+
+namespace FSH.WebApi.Infrastructure.Identity;
+
+internal partial class UserService
+{
+    public async Task<List<UserRoleDto>> GetRolesAsync(string userId, CancellationToken cancellationToken)
+    {
+        var userRoles = new List<UserRoleDto>();
+
+        var user = await _userManager.FindByIdAsync(userId);
+        var roles = await _roleManager.Roles.AsNoTracking().ToListAsync(cancellationToken);
+        foreach (var role in roles)
+        {
+            userRoles.Add(new UserRoleDto
+            {
+                RoleId = role.Id,
+                RoleName = role.Name,
+                Description = role.Description,
+                Enabled = await _userManager.IsInRoleAsync(user, role.Name)
+            });
+        }
+
+        return userRoles;
+    }
+
+    public async Task<string> AssignRolesAsync(string userId, UserRolesRequest request, CancellationToken cancellationToken)
+    {
+        ArgumentNullException.ThrowIfNull(request, nameof(request));
+
+        var user = await _userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync(cancellationToken);
+
+        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
+
+        // Check if the user is an admin for which the admin role is getting disabled
+        if (await _userManager.IsInRoleAsync(user, FSHRoles.Admin)
+            && request.UserRoles.Any(a => !a.Enabled && a.RoleName == FSHRoles.Admin))
+        {
+            // Get count of users in Admin Role
+            int adminCount = (await _userManager.GetUsersInRoleAsync(FSHRoles.Admin)).Count();
+
+            // Check if user is not Root Tenant Admin
+            // Edge Case : there are chances for other tenants to have users with the same email as that of Root Tenant Admin. Probably can add a check while User Registration
+            if (user.Email == MultitenancyConstants.Root.EmailAddress)
+            {
+                if (_currentTenant.Id == MultitenancyConstants.Root.Id)
+                {
+                    throw new ConflictException(_localizer["Cannot Remove Admin Role From Root Tenant Admin."]);
+                }
+            }
+            else if (adminCount <= 2)
+            {
+                throw new ConflictException(_localizer["Tenant should have at least 2 Admins."]);
+            }
+        }
+
+        foreach (var userRole in request.UserRoles)
+        {
+            // Check if Role Exists
+            if (await _roleManager.FindByNameAsync(userRole.RoleName) is not null)
+            {
+                if (userRole.Enabled)
+                {
+                    if (!await _userManager.IsInRoleAsync(user, userRole.RoleName))
+                    {
+                        await _userManager.AddToRoleAsync(user, userRole.RoleName);
+                    }
+                }
+                else
+                {
+                    await _userManager.RemoveFromRoleAsync(user, userRole.RoleName);
+                }
+            }
+        }
+
+        await _eventService.PublishAsync(new ApplicationUserUpdatedEvent(user.Id, true));
+
+        return _localizer["User Roles Updated Successfully."];
+    }
+}
