+++ src/Infrastructure/Identity/UserService.cs
@@ -4,49 +4,71 @@ using Finbuckle.MultiTenant;
 using FSH.WebApi.Application.Common.Caching;
 using FSH.WebApi.Application.Common.Events;
 using FSH.WebApi.Application.Common.Exceptions;
+using FSH.WebApi.Application.Common.FileStorage;
+using FSH.WebApi.Application.Common.Interfaces;
+using FSH.WebApi.Application.Common.Mailing;
 using FSH.WebApi.Application.Common.Models;
 using FSH.WebApi.Application.Common.Specification;
 using FSH.WebApi.Application.Identity.Users;
 using FSH.WebApi.Domain.Identity;
+using FSH.WebApi.Infrastructure.Mailing;
 using FSH.WebApi.Infrastructure.Persistence.Context;
 using FSH.WebApi.Shared.Authorization;
-using FSH.WebApi.Shared.Multitenancy;
 using Mapster;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Localization;
+using Microsoft.Extensions.Options;
 
 namespace FSH.WebApi.Infrastructure.Identity;
 
-internal class UserService : IUserService
+internal partial class UserService : IUserService
 {
+    private readonly SignInManager<ApplicationUser> _signInManager;
     private readonly UserManager<ApplicationUser> _userManager;
     private readonly RoleManager<ApplicationRole> _roleManager;
     private readonly ApplicationDbContext _db;
     private readonly IStringLocalizer<UserService> _localizer;
-    private readonly ITenantInfo _currentTenant;
+    private readonly IJobService _jobService;
+    private readonly IMailService _mailService;
+    private readonly MailSettings _mailSettings;
+    private readonly IEmailTemplateService _templateService;
+    private readonly IFileStorageService _fileStorage;
     private readonly IEventService _eventService;
     private readonly ICacheService _cache;
     private readonly ICacheKeyService _cacheKeys;
+    private readonly ITenantInfo _currentTenant;
 
     public UserService(
+        SignInManager<ApplicationUser> signInManager,
         UserManager<ApplicationUser> userManager,
         RoleManager<ApplicationRole> roleManager,
         ApplicationDbContext db,
         IStringLocalizer<UserService> localizer,
-        ITenantInfo currentTenant,
+        IJobService jobService,
+        IMailService mailService,
+        IOptions<MailSettings> mailSettings,
+        IEmailTemplateService templateService,
+        IFileStorageService fileStorage,
         IEventService eventService,
         ICacheService cache,
-        ICacheKeyService cacheKeys)
+        ICacheKeyService cacheKeys,
+        ITenantInfo currentTenant)
     {
+        _signInManager = signInManager;
         _userManager = userManager;
         _roleManager = roleManager;
         _db = db;
         _localizer = localizer;
-        _currentTenant = currentTenant;
+        _jobService = jobService;
+        _mailService = mailService;
+        _mailSettings = mailSettings.Value;
+        _templateService = templateService;
+        _fileStorage = fileStorage;
         _eventService = eventService;
         _cache = cache;
         _cacheKeys = cacheKeys;
+        _currentTenant = currentTenant;
     }
 
     public async Task<PaginationResponse<UserDetailsDto>> SearchAsync(UserListFilter filter, CancellationToken cancellationToken)
@@ -92,114 +114,6 @@ internal class UserService : IUserService
 
         return user.Adapt<UserDetailsDto>();
     }
-
-    public async Task<List<UserRoleDto>> GetRolesAsync(string userId, CancellationToken cancellationToken)
-    {
-        var userRoles = new List<UserRoleDto>();
-
-        var user = await _userManager.FindByIdAsync(userId);
-        var roles = await _roleManager.Roles.AsNoTracking().ToListAsync(cancellationToken);
-        foreach (var role in roles)
-        {
-            userRoles.Add(new UserRoleDto
-            {
-                RoleId = role.Id,
-                RoleName = role.Name,
-                Description = role.Description,
-                Enabled = await _userManager.IsInRoleAsync(user, role.Name)
-            });
-        }
-
-        return userRoles;
-    }
-
-    public async Task<string> AssignRolesAsync(string userId, UserRolesRequest request, CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(request, nameof(request));
-
-        var user = await _userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync(cancellationToken);
-
-        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
-
-        // Check if the user is an admin for which the admin role is getting disabled
-        if (await _userManager.IsInRoleAsync(user, FSHRoles.Admin)
-            && request.UserRoles.Any(a => !a.Enabled && a.RoleName == FSHRoles.Admin))
-        {
-            // Get count of users in Admin Role
-            int adminCount = (await _userManager.GetUsersInRoleAsync(FSHRoles.Admin)).Count();
-
-            // Check if user is not Root Tenant Admin
-            // Edge Case : there are chances for other tenants to have users with the same email as that of Root Tenant Admin. Probably can add a check while User Registration
-            if (user.Email == MultitenancyConstants.Root.EmailAddress)
-            {
-                if (_currentTenant.Id == MultitenancyConstants.Root.Id)
-                {
-                    throw new ConflictException(_localizer["Cannot Remove Admin Role From Root Tenant Admin."]);
-                }
-            }
-            else if (adminCount <= 2)
-            {
-                throw new ConflictException(_localizer["Tenant should have at least 2 Admins."]);
-            }
-        }
-
-        foreach (var userRole in request.UserRoles)
-        {
-            // Check if Role Exists
-            if (await _roleManager.FindByNameAsync(userRole.RoleName) is not null)
-            {
-                if (userRole.Enabled)
-                {
-                    if (!await _userManager.IsInRoleAsync(user, userRole.RoleName))
-                    {
-                        await _userManager.AddToRoleAsync(user, userRole.RoleName);
-                    }
-                }
-                else
-                {
-                    await _userManager.RemoveFromRoleAsync(user, userRole.RoleName);
-                }
-            }
-        }
-
-        await _eventService.PublishAsync(new ApplicationUserUpdatedEvent(user.Id, true));
-
-        return _localizer["User Roles Updated Successfully."];
-    }
-
-    public async Task<List<string>> GetPermissionsAsync(string userId, CancellationToken cancellationToken)
-    {
-        var user = await _userManager.FindByIdAsync(userId);
-
-        _ = user ?? throw new NotFoundException(_localizer["User Not Found."]);
-
-        var userRoles = await _userManager.GetRolesAsync(user);
-        var permissions = new List<string>();
-        foreach (var role in await _roleManager.Roles
-            .Where(r => userRoles.Contains(r.Name))
-            .ToListAsync(cancellationToken))
-        {
-            permissions.AddRange(await _db.RoleClaims
-                .Where(rc => rc.RoleId == role.Id && rc.ClaimType == FSHClaims.Permission)
-                .Select(rc => rc.ClaimValue)
-                .ToListAsync(cancellationToken));
-        }
-
-        return permissions.Distinct().ToList();
-    }
-
-    public async Task<bool> HasPermissionAsync(string userId, string permission, CancellationToken cancellationToken)
-    {
-        var permissions = await _cache.GetOrSetAsync(
-            _cacheKeys.GetCacheKey(FSHClaims.Permission, userId),
-            () => GetPermissionsAsync(userId, cancellationToken),
-            cancellationToken: cancellationToken);
-
-        return permissions?.Contains(permission) ?? false;
-    }
-
-    public Task InvalidatePermissionCacheAsync(string userId, CancellationToken cancellationToken) =>
-        _cache.RemoveAsync(_cacheKeys.GetCacheKey(FSHClaims.Permission, userId), cancellationToken);
 
     public async Task ToggleStatusAsync(ToggleUserStatusRequest request, CancellationToken cancellationToken)
     {
